pyproject.toml
[tool.poetry]
name = "atlatl"
version = "0.1.0"
description = ""
authors = ["heislandmine <heislandmine@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
"Mastodon.py" = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^20.8b1", allow-prereleases = true}
flake8 = "^3.8.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
